requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2-binary==2.8.6
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.3
